requirements.txt
streamlit
python-docx
openai
PyPDF2
openpyxl
xlrd
pandas
docx2txt
PyMuPDF